requirements.txt
fastapi
uvicorn
numpy
imageio
selenium
webdriver-manager
beautifulsoup4
requests
python-dotenv
